backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
google-generativeai
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
psutil
redis>=5.0.0
fastapi-limiter>=0.1.6
minio>=7.2.0
